Add animation of all letters in Arabic and Hebrew alphabets
This commit adds code to create an animation of all the letters in the Arabic and Hebrew alphabets. The code generates a DataFrame of glyphs for the letters, creates a folder for the output files if it doesn't exist, and saves the plots for each letter in the folder. The animation is created by combining the images of each letter in the folder.

The main function has been updated to call the new code and generate the animation. The code is also properly formatted and commented for clarity.